move logging related funcs to
